Move local group support folder to main module
