+++ rebel_management_utilities/resources/signups_message.md
