+++ rebel_management_utilities/utils/action_network.py
@@ -5,7 +5,7 @@ import requests
 from dotenv import load_dotenv
 from filecache import filecache
 
-from local_group_support.config.config import get_config
+from rebel_management_utilities.config.config import get_config
 
 API_URL = 'https://actionnetwork.org/api/v2/'
