+++ rebel_management_utilities/utils/nextcloud.py
@@ -3,7 +3,7 @@ import os
 import requests
 from dotenv import load_dotenv
 
-from local_group_support.utils.excel import append_df_to_excel
+from rebel_management_utilities.utils.excel import append_df_to_excel
 
 BASE_URL = 'https://cloud.extinctionrebellion.nl/remote.php/dav/files/'
 INTEGRATION_DIRECTORY = '/CloudXRNL/AppSpecific/Integrators_AN_Home/Integrators_FromLGs/'
